corrections

+++ app/Http/Livewire/Editprofile.php
@@ -7,16 +7,24 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Livewire\Component;
 use Livewire\WithFileUploads;
+use Illuminate\Support\Str;
+use Intervention\Image\ImageManagerStatic;
 
 class Editprofile extends Component
 {
+    use WithFileUploads;
     public $name;
     public $email;
     public $password;
     public $avatar;
-    use WithFileUploads;
+    public User $user;
 
-    public function profileEdit(){
+    public function mount() 
+    {
+        $this->user = Auth::user();
+    }
+
+    public function update(){
         $this->validate( [
             'name'=>'required',
             'email'=>'required|email',
@@ -26,11 +34,15 @@ class Editprofile extends Component
         $user->name = $this->name;
         $user->email = $this->email;
         $user->password = $this->password;
-        Storage::disk('public')->delete($user->avatar);
-        $user->avatar = Storage::disk( 'public' )->put( '', $this->avatar );
+        if ($user->avatar != null) {
+            Storage::disk('public')->delete($user->avatar);
+            $user->avatar = Storage::disk( 'public' )->put( '', $this->avatar );
+        }
         $user->save();
-        return redirect( route( 'dashboard' ) );
+        session()->flash('message', 'Profile successfully updated.');
+        return redirect()->to('/dashboard');
     }
+    
 
     public function render()
     {
